made photo index page generic
