+++ pages/travel/uzbekistan/photos/p/[photoId].tsx
@@ -1,18 +1,15 @@
-import { createGetStaticPaths } from '@src/photos/createGetStaticPaths'
-import { createGetStaticProps } from '@src/photos/createGetStaticProps'
+import { createGetStaticPaths } from '@src/photos/PhotoPageGetStaticPaths'
+import { createGetStaticProps } from '@src/photos/PhotoPageGetStaticProps'
 import { PhotoPage } from '@src/photos/PhotoPage'
 import type { ImageProps } from '@src/photos/utils/types'
 import type { GetStaticProps } from 'next'
-import { FC } from 'react'
 import { folderName } from '..'
 
 type Props = { currentPhoto: ImageProps }
 
-const Page: FC<Props> = ({ currentPhoto }: Props) => {
+export default function Page({ currentPhoto }: Props) {
     return <PhotoPage folderName={folderName} currentPhoto={currentPhoto} />
 }
-
-export default Page
 
 export const getStaticProps: GetStaticProps = async (context) => {
     return createGetStaticProps(folderName, context)
