+++ src/photos/PhotoPageGetStaticPaths.ts
