+++ src/photos/PhotoPageGetStaticProps.ts
